+++ lib/content/mod.ts
@@ -1,4 +1,5 @@
 export * from "./model.ts";
 export * from "./asset.ts";
 export * from "./interfaces.ts";
+export * from "./generators.ts";
 export * from "./md-page.ts";